refactor: 댓글 조회 outbound 포트 어댑터 의존 제거 및 도메인 모델 이동
