+++ src/main/java/com/dreamteam/alter/domain/workspace/model/WorkspaceRequestCommentListResponse.java
@@ -1,0 +1,22 @@
+package com.dreamteam.alter.domain.workspace.model;
+
+import java.time.LocalDateTime;
+
+import com.dreamteam.alter.domain.workspace.type.CommentOwner;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class WorkspaceRequestCommentListResponse {
+
+	private Long id;
+	private Long workspaceRequestId;
+	private Long userId;
+	private CommentOwner commentOwner;
+	private String comment;
+	private LocalDateTime createdAt;
+}
