+++ src/main/java/com/dreamteam/alter/domain/workspace/port/outbound/WorkspaceRequestCommentQueryRepository.java
@@ -2,7 +2,7 @@ package com.dreamteam.alter.domain.workspace.port.outbound;
 
 import java.util.List;
 
-import com.dreamteam.alter.adapter.outbound.workspace.persistence.readonly.WorkspaceRequestCommentListResponse;
+import com.dreamteam.alter.domain.workspace.model.WorkspaceRequestCommentListResponse;
 
 public interface WorkspaceRequestCommentQueryRepository {
 	List<WorkspaceRequestCommentListResponse> getCommentList(Long workspaceRequestId);
